+++ Source/OxyPlot/Series/BarSeries/BarSeries.cs
@@ -9,6 +9,7 @@
 
 namespace OxyPlot.Series
 {
+    using OxyPlot.Axes;
     using System;
     using System.Collections.Generic;
     using System.Linq;
@@ -37,6 +38,7 @@ namespace OxyPlot.Series
             this.NegativeFillColor = OxyColors.Undefined;
             this.TrackerFormatString = DefaultTrackerFormatString;
             this.LabelMargin = 2;
+            this.LabelAngle = 0;
             this.StackGroup = string.Empty;
             this.StrokeThickness = 0;
         }
@@ -46,7 +48,7 @@ namespace OxyPlot.Series
         /// </summary>
         /// <value>The actual color.</value>
         public OxyColor ActualFillColor => this.FillColor.GetActualColor(this.defaultFillColor);
-
+        
         /// <summary>
         /// Gets or sets the base value.
         /// </summary>
@@ -77,9 +79,14 @@ namespace OxyPlot.Series
         public string LabelFormatString { get; set; }
 
         /// <summary>
-        /// Gets or sets the label margins.
+        /// Gets or sets the label margins. Default value is 2.
         /// </summary>
         public double LabelMargin { get; set; }
+
+        /// <summary>
+        /// Gets or sets the label angle in degrees. Default value is 0.
+        /// </summary>
+        public double LabelAngle { get; set; }
 
         /// <summary>
         /// Gets or sets label placements.
@@ -307,40 +314,43 @@ namespace OxyPlot.Series
             double categoryEndValue)
         {
             var s = StringHelper.Format(this.ActualCulture, this.LabelFormatString, item, item.Value);
-            HorizontalAlignment ha;
             ScreenPoint pt;
             var y = (categoryEndValue + categoryValue) / 2;
             var sign = Math.Sign(topValue - baseValue);
             var marginVector = new ScreenVector(this.LabelMargin, 0) * sign;
+            var centreVector = new ScreenVector(0, 0);
+
+            var size = rc.MeasureText(
+                s,
+                this.ActualFont,
+                this.ActualFontSize,
+                this.ActualFontWeight,
+                this.LabelAngle);
 
             switch (this.LabelPlacement)
             {
                 case LabelPlacement.Inside:
                     pt = this.Transform(topValue, y);
                     marginVector = -marginVector;
-                    ha = (HorizontalAlignment)sign;
+                    centreVector = new ScreenVector(-sign * size.Width / 2, 0);
                     break;
                 case LabelPlacement.Outside:
                     pt = this.Transform(topValue, y);
-                    ha = (HorizontalAlignment)(-sign);
+                    centreVector = new ScreenVector(sign * size.Width / 2, 0);
                     break;
                 case LabelPlacement.Middle:
                     pt = this.Transform((topValue + baseValue) / 2, y);
                     marginVector = new ScreenVector(0, 0);
-                    ha = HorizontalAlignment.Center;
                     break;
                 case LabelPlacement.Base:
                     pt = this.Transform(baseValue, y);
-                    ha = (HorizontalAlignment)(-sign);
+                    centreVector = new ScreenVector(sign * size.Width / 2, 0);
                     break;
                 default:
                     throw new ArgumentOutOfRangeException();
             }
 
-            var va = VerticalAlignment.Middle;
-            this.Orientate(ref ha, ref va);
-
-            pt += this.Orientate(marginVector);
+            pt += this.Orientate(marginVector) + this.Orientate(centreVector);
 
             rc.DrawText(
                 pt,
@@ -349,9 +359,9 @@ namespace OxyPlot.Series
                 this.ActualFont,
                 this.ActualFontSize,
                 this.ActualFontWeight,
-                0,
-                ha,
-                va);
+                this.LabelAngle,
+                HorizontalAlignment.Center,
+                VerticalAlignment.Middle);
         }
 
         /// <inheritdoc/>
@@ -387,6 +397,11 @@ namespace OxyPlot.Series
 
                 var topValue = this.IsStacked ? baseValue + value : value;
 
+                if (this.YAxis.IsLogarithmic() && !this.YAxis.IsValidValue(topValue))
+                {
+                    continue;
+                }
+
                 // Calculate offset
                 double categoryValue;
                 if (this.IsStacked)
@@ -401,6 +416,11 @@ namespace OxyPlot.Series
                 if (this.IsStacked)
                 {
                     this.Manager.SetCurrentBaseValue(stackIndex, categoryIndex, value < 0, topValue);
+                }
+
+                if (this.YAxis.IsLogarithmic() && !this.YAxis.IsValidValue(baseValue))
+                {
+                    baseValue = LogarithmicAxis.LowestValidRoundtripValue;
                 }
 
                 var rect = new OxyRect(this.Transform(baseValue, categoryValue), this.Transform(topValue, categoryValue + actualBarWidth));