+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Example to Show/Hide Legend (#1470)
 - Example of BarSeries stacked and with labels (#1979)
 - Example of issue with AreaSeries tracker (#1982)
+- BarSeries.LabelAngle property (#1870)
 
 ### Changed
 
@@ -18,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - SystemInvalidException in LineSeries when only Double.Nan values are added (#1991)
 - Issue with tracking AreaSeries with monotonic data points (#1982)
 - HitTest when IsLegendVisible is false (#1975)
+- Font weight not being applied in ImageSharp (#2006)
 
 ## [2.1.2] - 2022-12-03
 
@@ -48,6 +50,7 @@ All notable changes to this project will be documented in this file.
 - CategoryAxis.ItemsSource without BarSeries (#1847)
 - Histogram now rendering properly when using logarithmic Y axis (#740) 
 - Fix ExampleLibrary build errors in certain code pages (#1890)
+- LineBarSeries now rendering properly when using logarithmic scale (#740)
 - Fix for double.Epsilon zero check that fails on some architectures (#1924)
 
 ## [2.1.0] - 2021-10-02